🐛 [-bug] Run pytest from the source root (for pyproject.toml access)
Previously, pytest was run from the "tests" directory. But pytest did not read settings from pyproject.toml when run from that directory. With this change, tests are run from the root allowing access to settings in pyproject.toml.

Changes were also made to the CI workflows to upload .coverage files if they were found at the root or in the tests directory.